fixed linting issues + addressed changes requested in PR
