+++ ALUM/ALUM/Components/SignUpGradeComponent.swift
@@ -8,17 +8,17 @@
 import SwiftUI
 
 struct SignUpGradeComponent: View {
-    
+
     var grade: String
     var isSelected: Bool
-    
+
     var body: some View {
         HStack {
             ZStack {
                 Circle()
                     .strokeBorder(Color("ALUM Dark Blue"), lineWidth: 2.0)
                     .frame(width: 20, height: 20)
-                
+
                 if isSelected {
                     Circle()
                         .fill(Color("ALUM Dark Blue"))
@@ -29,7 +29,7 @@ struct SignUpGradeComponent: View {
             .padding(.trailing, 10)
             .padding(.top, 14)
             .padding(.bottom, 14)
-            
+
             Text(grade)
                 .font(.custom("Metropolis-Regular", size: 17))
                 .padding(.trailing, 20)
@@ -41,7 +41,7 @@ struct SignUpGradeComponent: View {
                     .padding(.trailing, 8)
                     .padding(.leading, 8)
                     .frame(height: 48.0)
-                
+
                 RoundedRectangle(cornerRadius: 12.0)
                     .stroke(Color("ALUM Light Blue"))
                     .padding(.trailing, 8)
@@ -52,10 +52,12 @@ struct SignUpGradeComponent: View {
     }
 }
 
-/*
 struct SignUpGradeComponent_Previews: PreviewProvider {
+
+    static var grade: String = "9"
+    static var isSelected: Bool = true
+
     static var previews: some View {
-        SignUpGradeComponent()
+        SignUpGradeComponent(grade: grade, isSelected: isSelected)
     }
 }
-*/
