+++ ALUM/ALUM/Views/LoginPage/LoginPageView.swift
@@ -86,11 +86,11 @@ struct LoginPageView: View {
                 .padding(.trailing, 16)
                 .padding(.bottom, 32)
             }
-            
+
             HStack {
                 Text("Don't have an account?")
                     .font(.custom("Metropolis-Regular", size: 16))
-                
+
                 NavigationLink(destination: SignUpPageView(), label: {
                     Text("Sign-Up")
                         .underline()
